Only enable on main branch

+++ enterprise/dev/ci/ci/pipeline-steps.go
@@ -167,15 +167,21 @@ func addDockerfileLint(pipeline *bk.Pipeline) {
 }
 
 // Adds backend integration tests step.
-func addBackendIntegrationTests(pipeline *bk.Pipeline) {
-	pipeline.AddStep(":chains:",
-		bk.Cmd("pushd enterprise"),
-		bk.Cmd("./cmd/server/pre-build.sh"),
-		bk.Cmd("./cmd/server/build.sh"),
-		bk.Cmd("popd"),
-		bk.Cmd("./dev/ci/backend-integration.sh"),
-		bk.Cmd(`docker image rm -f "$IMAGE"`),
-	)
+func addBackendIntegrationTests(c Config) func(*bk.Pipeline) {
+	return func(pipeline *bk.Pipeline) {
+		if !c.isMasterDryRun && c.branch != "master" && c.branch != "main" {
+			return
+		}
+
+		pipeline.AddStep(":chains:",
+			bk.Cmd("pushd enterprise"),
+			bk.Cmd("./cmd/server/pre-build.sh"),
+			bk.Cmd("./cmd/server/build.sh"),
+			bk.Cmd("popd"),
+			bk.Cmd("./dev/ci/backend-integration.sh"),
+			bk.Cmd(`docker image rm -f "$IMAGE"`),
+		)
+	}
 }
 
 func addBrowserExtensionE2ESteps(pipeline *bk.Pipeline) {
